Fix TPU initialization for local servers
Requires an identity in the TPU initialization function to avoid placement errors. I believe this only comes up when using local servers (i.e. affects mostly testing; we do have plenty of tests for TPUs on remote jobs).

Also exposes a mapping from job name to TPU topology. I have a use for it: we need to look up the topology corresponding to the correct job when replicating a function.

PiperOrigin-RevId: 290188617
Change-Id: I24e1e5995f6f55b565a1aac05909698ac3ee49d8