Use resource subtype if present in DecomposeResourceOps pass.
When a resource subtype is present, the output of the generated tf.ReadVariableOp after decomposing a resource op can be set to the subtype. Otherwise, an unranked tensor of element type is used as the output type instead. This removes some potential dynamic shapes that shape inference does not handle currently.

PiperOrigin-RevId: 290749517
Change-Id: I97809a067f876225daf830514973d743423a983a