Add TFL while op with regions
Add TFL while op that uses regions for the conditional and body of the loop.
This op has the same semantics as TF while but uses regions instead. This is
just the smallest start to using the TFL while op:
* the LoopLike interface as the moment is sparse and needs to be filled in;
* the legalization from TF while to TFL while is also pending (and so currently
  supports both forms as legal during translation to flatbuffer);
* extra verification is needed for the regions of the while;
* missing pretty form;
* import from flatbuffer will not generate this form;

The same export test is used for tfl.while as for tf.While op (modulo the
names).

PiperOrigin-RevId: 292557025
Change-Id: I3b1315361a51706273594aba6e0c6fe0fc321869