Refactor sorting logic of PrioritizedDeviceTypeVector.
We will need to construct a PrioritizedDeviceTypeVector from a
DeviceSet inside ConvertToEagerTensor in order to fix b/139690309
(because we will need to call Context::SelectDevice).

The natural place to have that list created is DeviceSet, and once
that can be done from there, it is also natural to move the sorting
logic there, reusing it from ColocationGraph.

It also makes sense to compute it only once, since its value does not
change once a DeviceSet is built.

It also made sense to create a similar PrioritizedDeviceVector at this
time, and use it ColocationGraph. It will also be used in changes to come.

PiperOrigin-RevId: 292567038
Change-Id: I8ee9ef0ca16d390d32fdcf25c2447827dc0d1d53