[XLA:Python] Add support for zero-copy NumPy -> XLA buffer transfers on CPU.
When creating a PyLocalBuffer from a NumPy array, we don't need to copy the array if it is in the right layout and alignment already. This saves us a buffer copy and some thread hops.

Note that a copy still occurs if the input NumPy array was not in C-contiguous layout. A copy will also take place if the NumPy array was not 64-byte aligned.

Remove support for tuple trees in Buffer.from_pyval. JAX doesn't use tuple trees, and this simplifies the code and the API.

PiperOrigin-RevId: 292650189
Change-Id: If6f7b34507c0aebc6dcbee136049b66e66d426c3