Make ShapeIsCompatible handle dynamic dimensions.
Our codegen will always generate loops over the full shape dimensions, so the
index can be used as long as the static shape dimensions have the same size.
If that were to change, we would have to track also whether a dimension was
dynamic when we create the IrArray::Index, and what the real size is.

PiperOrigin-RevId: 292906320
Change-Id: I930f9ec336bcf7eca97d03a3fe798bc778075a3a