Depthwise convolution 3x3 per-channel int8 for dot-product ARM (8).
Generalize kernel intrinsics routines for symmetric int8.

For overall simplicity and commonality of code this updates the kernels to be less specific to the data type. The idea is that the main difference between the kernels will be the per-channel aspect and not uint8 vs symmetric int8. This is not fully achieved, since the code can become unduly complicated.

Sadly uint-8x16, int-8x16 data types are the same to the toolchain (implemented as macros) in some cases, and the resulting lack of signature difference means that the min and max functions cannot be commonly templated across all the routines as we did with load and store.

PiperOrigin-RevId: 292919343
Change-Id: Iec12c383571e6b983c654beab5180a56a0b0ce4d