Do not fail when TPUReplicateMetadata op is missing.
The rationale behind this change is that functionalizing control flow in TF v1 models adds functions to function library with nodes that have '_tpu_replicate' attrribute. The TPU cluster formation pass throws error on these functions when it does not find the TPUReplicateMetadata op associated with a _tpu_replicate attribute.

PiperOrigin-RevId: 292986754
Change-Id: I203c4e62db96c835bdad3f669500674dc4fce8c5