Add new_axis_mask support to `tf.StridedSlice` lowering.
To add new_axis_mask support we only need to keep track for the number of new axis after ellipsis. This helps in counting the number of dimensions ellipsis represents and hence produce the correct attributes of  `xla_hlo.slice` op. The result of slice is reshaped to reflect new axes added.

PiperOrigin-RevId: 294342718
Change-Id: Ic2fc3d9116fd2160a5341826180b1b0bdc231698