cuda_configure: support toolchain generation where the host and execution platform are different
I discovered that str(path("C:\\foo")) == "C:\foo" does not always hold. In particular if the host platform is Linux an the execution platform is Windows. That is because path() only knows about the host() platform and its internal representation is absolute paths. So on Linux C:\foo looks like a relative path to path() and it will turn it into an absolute path prefixed with the external repository directory. With this change it's possible to create a cuda configuration for a remote Windows platform from a Linux laptop.

PiperOrigin-RevId: 294705239
Change-Id: I37c931961db45d410cde115d8a2158fc48a626d0