[SparseToDense] Fix benchmark reporting for the SparseToDense op.
Previously, we were using a concatenated benchmark arg for the size and rank, which appears to have led to name collisions when the benchmark reporter makes the benchmark name human readable. This change switches the benchmark to using an explicit `ArgPair` for those arguments, which reports them separately.

In addition, this change switches the benchmark to use a DT_INT64 indices tensor, which matches the common case for tf.SparseTensor usage in the TensorFlow API, and avoids a cast in the SparseToDenseOp implementation.

PiperOrigin-RevId: 295985981
Change-Id: I94df39e2ef64b387f4c158ef0c8b33153c3cca4c