Fix the cache key problem when compute_output_shape().
This is a very tricky one wrt the id() of int in python. Under the hood,
id returns memory address for the int, and python has a cache location
for the ints, which result into different ints get same hash value.

Changed to use tuples of shape itself as the dict key, since the tuple
itself is immutable and hashable. Same tuple value will return the same
hash value.

Also remove the generic utils for that where network.py is only usage
for that function.

Fix #32029

PiperOrigin-RevId: 296302946
Change-Id: I865c9380a06ed6ee80fea7f942c21c4d102473c2